Update dependencies Use pnpm instead of Yarn Add webpack as devDependency Actualize ESLint/prettier config Add .editorconfig
