+++ package.json
@@ -26,43 +26,37 @@
     "lint-staged": "lint-staged"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "2.34.0",
-    "@typescript-eslint/parser": "2.34.0",
-    "eslint": "7.0.0",
-    "eslint-config-prettier": "6.11.0",
-    "eslint-formatter-pretty": "3.0.1",
-    "eslint-plugin-import": "2.20.2",
-    "eslint-plugin-import-helpers": "1.0.2",
-    "eslint-plugin-prettier": "3.1.3",
-    "husky": "4.2.5",
-    "lint-staged": "10.2.4",
-    "npm-run-all": "4.1.5",
-    "prettier": "2.0.5",
-    "typescript": "3.9.3",
-    "@types/webpack": "4.41.13",
-    "@types/node": "14.0.4"
+    "@typescript-eslint/eslint-plugin": "^4.8.2",
+    "@typescript-eslint/parser": "^4.8.2",
+    "eslint": "^7.14.0",
+    "eslint-config-prettier": "^6.15.0",
+    "eslint-formatter-pretty": "^4.0.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-import-helpers": "^1.1.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "husky": "^4.3.0",
+    "lint-staged": "^10.5.2",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.2.0",
+    "typescript": "^4.1.2",
+    "webpack": "^4.44.2",
+    "@types/webpack": "^4.41.25",
+    "@types/node": "^14.14.10"
   },
   "peerDependencies": {
-    "webpack": "^4.0.0"
-  },
-  "peerDependenciesMeta": {
-    "webpack": {
-      "optional": true
-    }
+    "webpack": "4.x"
   },
   "husky": {
     "hooks": {
-      "pre-commit": "run-s lint-staged"
+      "pre-commit": "pnpx lint-staged"
     }
   },
   "lint-staged": {
-    "*.{js,ts}": [
-      "eslint --format=pretty --fix",
-      "git add"
+    "*.{js,ts,jsx,tsx,md,json}": [
+      "prettier --write"
     ],
-    "*.{json,md}": [
-      "prettier --write",
-      "git add"
+    "*.{js,ts,jsx,tsx}": [
+      "eslint --format=pretty --fix"
     ]
   },
   "keywords": [
